+++ Sources/NotepadMacCore/SessionState.swift
@@ -47,33 +47,59 @@ public struct EditorWindowSessionState: Codable, Equatable {
 public struct EditorSessionState: Codable, Equatable {
     public let id: String
     public let filePath: String?
-    public let text: String
-    public let originalText: String
     public let selectedLocation: Int
     public let wordWrapEnabled: Bool
     public let statusBarVisible: Bool
     public let zoomPercent: Int
     public let lineEnding: LineEnding
 
+    private enum CodingKeys: String, CodingKey {
+        case id
+        case filePath
+        case selectedLocation
+        case wordWrapEnabled
+        case statusBarVisible
+        case zoomPercent
+        case lineEnding
+    }
+
     public init(
         id: String,
         filePath: String?,
-        text: String,
-        originalText: String,
         selectedLocation: Int,
         wordWrapEnabled: Bool,
         statusBarVisible: Bool,
         zoomPercent: Int,
-        lineEnding: LineEnding = .windows
+        lineEnding: LineEnding
     ) {
         self.id = id
         self.filePath = filePath
-        self.text = text
-        self.originalText = originalText
         self.selectedLocation = selectedLocation
         self.wordWrapEnabled = wordWrapEnabled
         self.statusBarVisible = statusBarVisible
         self.zoomPercent = zoomPercent
         self.lineEnding = lineEnding
+    }
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        id = try container.decode(String.self, forKey: .id)
+        filePath = try container.decodeIfPresent(String.self, forKey: .filePath)
+        selectedLocation = try container.decodeIfPresent(Int.self, forKey: .selectedLocation) ?? 0
+        wordWrapEnabled = try container.decodeIfPresent(Bool.self, forKey: .wordWrapEnabled) ?? true
+        statusBarVisible = try container.decodeIfPresent(Bool.self, forKey: .statusBarVisible) ?? true
+        zoomPercent = try container.decodeIfPresent(Int.self, forKey: .zoomPercent) ?? 100
+        lineEnding = try container.decodeIfPresent(LineEnding.self, forKey: .lineEnding) ?? .windows
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(id, forKey: .id)
+        try container.encodeIfPresent(filePath, forKey: .filePath)
+        try container.encode(selectedLocation, forKey: .selectedLocation)
+        try container.encode(wordWrapEnabled, forKey: .wordWrapEnabled)
+        try container.encode(statusBarVisible, forKey: .statusBarVisible)
+        try container.encode(zoomPercent, forKey: .zoomPercent)
+        try container.encode(lineEnding, forKey: .lineEnding)
     }
 }